:recycle: 라우터 상수 데이터로 분리
